+++ cmd/commands/root.go
@@ -13,7 +13,7 @@ import (
 var (
 	cfgFile  string
 	gitToken string
-	Version  = "3.9.0"
+	Version  = "3.10.0"
 )
 
 var RootCmd = &cobra.Command{